refactor: update module imports to use .js extension and adjust paths
- Changed main entry point in package.json from './dist/src/index.js' to './dist/index.js'.
- Updated development script in package.json to reference 'index.ts' directly.
- Modified import statements in various files to include the .js extension for consistency.
- Created a new app.ts file for Express application setup with middleware and error handling.
- Adjusted database configuration imports to use .js extension.
- Updated authentication feature files to reflect new import paths and .js extensions.
- Refactored validation schemas and utility functions to ensure proper module resolution.
- Updated tsconfig.json to extend from @tsconfig/node22 and simplified configuration.
