+++ package.json
@@ -20,10 +20,10 @@
   "repository": "https://github.com/factman/WalletApi.git",
   "homepage": "https://github.com/factman/WalletApi#readme",
   "type": "module",
-  "main": "./dist/src/index.js",
+  "main": "./dist/index.js",
   "scripts": {
     "start": "node .",
-    "dev": "tsx --watch --env-file .env src/index.ts",
+    "dev": "tsx --watch --env-file .env index.ts",
     "build": "tsc -p tsconfig.build.json",
     "type-check": "tsc --noEmit",
     "lint": "eslint .",
@@ -59,6 +59,7 @@
   "devDependencies": {
     "@dotenvx/dotenvx": "^1.44.1",
     "@eslint/js": "^9.27.0",
+    "@tsconfig/node22": "^22.0.2",
     "@types/axios": "^0.9.36",
     "@types/cors": "^2.8.18",
     "@types/express": "^5.0.2",
