+++ src/helpers/__tests__/validations.spec.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from "vitest";
 import { z } from "zod";
 
-import { HeaderSchemaType } from "../types";
-import { buildHeaderSchema, buildStrictSchema } from "../validations";
+import { HeaderSchemaType } from "../types.js";
+import { buildHeaderSchema, buildStrictSchema } from "../validations.js";
 
 describe("buildStrictSchema", () => {
   it("should create a strict schema that only allows defined properties", () => {
