+++ src/repositories/AuthenticatedUserRepository.ts
@@ -1,8 +1,7 @@
-import database from "@/configs/database";
-import { SCHEMA_VIEWS } from "@/helpers/constants";
-import AuthenticatedUserModel from "@/models/AuthenticatedUserModel";
-
-import { Repository } from "./Repository";
+import database from "../configs/database.js";
+import { SCHEMA_VIEWS } from "../helpers/constants.js";
+import AuthenticatedUserModel from "../models/AuthenticatedUserModel.js";
+import { Repository } from "./Repository.js";
 
 export class AuthenticatedUserRepository extends Repository<AuthenticatedUserModel> {
   constructor(databaseInstance = database) {
