+++ src/repositories/SessionRepository.ts
@@ -1,10 +1,9 @@
 import Knex from "knex";
 
-import database from "@/configs/database";
-import { SCHEMA_TABLES } from "@/helpers/constants";
-import SessionModel from "@/models/SessionModel";
-
-import { Repository } from "./Repository";
+import database from "../configs/database.js";
+import { SCHEMA_TABLES } from "../helpers/constants.js";
+import SessionModel from "../models/SessionModel.js";
+import { Repository } from "./Repository.js";
 
 export class SessionRepository extends Repository<SessionModel> {
   constructor(databaseInstance = database) {
